Resolving Ticket #64. Put the group definitions into doc/doxygroups.txt
Removed all the code from the headers that create the group. hopefully this makes the headers prettier. 

After I did this I realized that I was wrong, there were defgroup's in a lot of the headers. 
These did not need to be there, and I removed them. For each group there only neeeds to be one defgroup for it. 
To place something in a group you just need to do @ingroup groupname. 

Anwyays, this should close #64.