Doxygend lots of ben's README's and comment blocks.
Documented the pipeline a bit. Added a media section to the
main page. 

Tweaked some doxygen.cfg paremeters. 

Added documentation to a lot of the programs. 
A description of each program now shows up on the doxygen page. 

Every namespace variable is documented at least with a @brief 
description. 

I'm going to say this Resolves #85 but everyone should take a look to make
sure they agree.