Removed entdone
The idea behind entdone was that you could perform a series of
modifications to the ents without worrying about the screendraw happening
in an inconsistent state.  The problem comes up when you want to runs
scripts at regular intervals:

	every 10 {
		c3 destroy
		ent c3 {circle 10 {10 10} {10 10 10}}
		entdone }

This needs an entdone or it's changes wont be accepted, but this entdone
could happen in the middle of a series if inputs from stdin.  Which defeats
the whole purpose of entdone.

To avoid this problem, we suggest that you simply wrap any code that needs
to happen all-at-once with eval:

	eval {
		shift c3
		shift c4 }

Which solves the problem because the Tcl event loop is not premtive.